temp commit for merging

+++ migrations/20160324120647-create-crime.js
@@ -23,10 +23,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       latitude: {
-        type: Sequelize.DECIMAL(10, 6)
+        type: Sequelize.FLOAT
       },
       longitude: {
-        type: Sequelize.DECIMAL(10, 6)
+        type: Sequelize.FLOAT
       },
       createdAt: {
         allowNull: false,
